Use builder pattern to avoid constructor magic.
As we are using a builder pattern already, we might as well just expose it.
This simplifies the constructor (just set internal final fields), relieves
the need for parameter juggling between builder and internal constructor,
and allows for checking incoming parameters *before* using them to construct
anything.

If this actually makes it into a public library, consider expanding builder
pattern to include both from/to and offset/len pattern? This should be easy.